Fixed an Idempotency Problem In Cumulus Bond Provider
The slave attribute on the provider wasn't being parsed properly due to cumulus nclu merging all of the slaves on to one line. The provider now passes the list through the extrapolate_list function to ensure that the cumulus abbreviated interface format is properly split into each of it's components.